+++ syndicate/core/constants.py
@@ -304,3 +304,7 @@ PARAMETER_NAME_PARAM = 'parameter_name'
 EMPTY_VALUES = [[], set(), (), {}, '']
 
 AUTHORIZATION_SCOPES_KEY = 'authorization_scopes'
+
+
+DEPLOYED_MARKER = '\u2713 Deployed'   # ✓ Deployed
+LOCAL_MARKER = '\u2717 Local'         # ✗ Local